Revert "Revert "Revert "Add manabase generator, hidden from the menu for now"""
This reverts commit 76c33dec45f6d4de4bb0c6f2b02cfe7579cf1800.
