--- decksite/templates/recentdecks.mustache
@@ -1,11 +1,0 @@
-{{#form.has_recent_decks}}
-    <div>
-        <label for="signup_recent_decks">{{TT_YOUR_RECENT_DECKS}}</label>
-        <select id="signup_recent_decks">
-            <option value="">{{TT_CHOOSE_DECK}}</option>
-            {{#form.recent_decks}}
-                <option value="" data="{{list}}">{{name}}</option>
-            {{/form.recent_decks}}
-        </select>
-    </div>
-{{/form.has_recent_decks}}
