+++ decksite/views/__init__.py
@@ -28,7 +28,6 @@ from .kickoff import KickOff
 from .league_form import LeagueForm
 from .league_info import LeagueInfo
 from .link_accounts import LinkAccounts
-from .manabase import Manabase
 from .matchups import Matchups
 from .metagame import Metagame
 from .pd500 import PD500
@@ -53,4 +52,4 @@ from .tournament_leaderboards import TournamentLeaderboards
 from .tournaments import Tournaments
 from .unlink import Unlink
 
-__all__ = ['About', 'AboutPdm', 'Achievements', 'Admin', 'AdminRetire', 'Archetype', 'Archetypes', 'Ban', 'Banners', 'Bugs', 'Card', 'Cards', 'CommunityGuidelines', 'Competition', 'Competitions', 'ContactUs', 'Deck', 'DeckEmbed', 'DeckCheck', 'Decks', 'EditAliases', 'EditArchetypes', 'EditLeague', 'EditMatches', 'EditRules', 'Faqs', 'Home', 'KickOff', 'LeagueForm', 'LeagueInfo', 'LinkAccounts', 'Manabase', 'Matchups', 'Metagame', 'PD500', 'People', 'Person', 'PersonAchievements', 'PersonMatches', 'PlayerNotes', 'Prizes', 'Report', 'Resources', 'Retire', 'Rotation', 'RotationChanges', 'RotationChecklist', 'RotationSpeculation', 'Seasons', 'SignUp', 'Sorters', 'TournamentHosting', 'TournamentLeaderboards', 'Tournaments', 'Unlink']
+__all__ = ['About', 'AboutPdm', 'Achievements', 'Admin', 'AdminRetire', 'Archetype', 'Archetypes', 'Ban', 'Banners', 'Bugs', 'Card', 'Cards', 'CommunityGuidelines', 'Competition', 'Competitions', 'ContactUs', 'Deck', 'DeckEmbed', 'DeckCheck', 'Decks', 'EditAliases', 'EditArchetypes', 'EditLeague', 'EditMatches', 'EditRules', 'Faqs', 'Home', 'KickOff', 'LeagueForm', 'LeagueInfo', 'LinkAccounts', 'Matchups', 'Metagame', 'PD500', 'People', 'Person', 'PersonAchievements', 'PersonMatches', 'PlayerNotes', 'Prizes', 'Report', 'Resources', 'Retire', 'Rotation', 'RotationChanges', 'RotationChecklist', 'RotationSpeculation', 'Seasons', 'SignUp', 'Sorters', 'TournamentHosting', 'TournamentLeaderboards', 'Tournaments', 'Unlink']
