+++ decksite/views/decklist_form.py
@@ -1,8 +1,6 @@
-from gettext import gettext
-
 from flask import url_for
 
-from decksite.form import Form
+from decksite.data.form import Form
 from decksite.views.league_form import LeagueForm
 
 
@@ -21,15 +19,3 @@ class DecklistForm(LeagueForm):
         if self.form.card_warnings is not None:
             self.has_warnings = 'Warnings_Bugs' in self.form.card_warnings and len(self.form.card_warnings['Warnings_Bugs']) > 0
             self.bugged_cards_url = url_for('tournaments', _anchor='bugs')
-
-    def TT_DECKLIST(self) -> str:
-        return gettext('Decklist')
-
-    def TT_ENTER_OR_UPLOAD(self) -> str:
-        return gettext('Enter or upload your decklist')
-
-    def TT_YOUR_RECENT_DECKS(self) -> str:
-        return gettext('Your Recent Decks')
-
-    def TT_CHOOSE_DECK(self) -> str:
-        return gettext('Select a recent deck to start from there')
