+++ decksite/views/report.py
@@ -1,7 +1,7 @@
 from flask import url_for
 from flask_babel import gettext
 
-from decksite.form import Form
+from decksite.data.form import Form
 from decksite.views.league_form import LeagueForm
