+++ decksite/views/signup.py
@@ -1,7 +1,7 @@
 from flask import url_for
 from flask_babel import gettext
 
-from decksite.form import Form
+from decksite.data.form import Form
 from decksite.views.decklist_form import DecklistForm
 from magic.models import Deck
 
@@ -24,5 +24,17 @@ class SignUp(DecklistForm):
     def TT_DECK_NAME(self) -> str:
         return gettext('Deck Name')
 
+    def TT_DECKLIST(self) -> str:
+        return gettext('Decklist')
+
+    def TT_ENTER_OR_UPLOAD(self) -> str:
+        return gettext('Enter or upload your decklist')
+
     def TT_SIGNUP(self) -> str:
         return gettext('Sign Up')
+
+    def TT_YOUR_RECENT_DECKS(self) -> str:
+        return gettext('Your Recent Decks')
+
+    def TT_CHOOSE_DECK(self) -> str:
+        return gettext('Select a recent deck to start from there')
